Moved AltersDataMixin from `template.utils` to `utils`
